refactor(desktop): split renderer App.tsx into feature modules
Extract setup-wizard/, hero/, command-log/, modals/, components.tsx, and
format.ts from the ~1,900-line renderer App.tsx. Pure code move; no
behavior change. App.test.tsx is untouched and still drives the real tree.

Closes #19839
